feat(3dtiles): rename ThreeDTilesLayer to OGC3DTilesLayer
