+++ examples/3dtiles_basic.html
@@ -45,7 +45,7 @@
 
             // Create a new Layer 3d-tiles For DiscreteLOD
             // -------------------------------------------
-            var $3dTilesLayerDiscreteLOD = new itowns.ThreeDTilesLayer('3d-tiles-discrete-lod', {
+            var $3dTilesLayerDiscreteLOD = new itowns.OGC3DTilesLayer('3d-tiles-discrete-lod', {
                 name: 'DiscreteLOD',
                 sseThreshold: 0.05,
                 source: new itowns.C3DTilesSource({
@@ -58,7 +58,7 @@
             // Create a new Layer 3d-tiles For Viewer Request Volume
             // -----------------------------------------------------
 
-            var $3dTilesLayerRequestVolume = new itowns.ThreeDTilesLayer('3d-tiles-request-volume', {
+            var $3dTilesLayerRequestVolume = new itowns.OGC3DTilesLayer('3d-tiles-request-volume', {
                 name: 'RequestVolume',
                 source: new itowns.C3DTilesSource({
                     url: 'https://raw.githubusercontent.com/CesiumGS/3d-tiles-samples/master/1.0/TilesetWithRequestVolume/tileset.json',
