+++ examples/3dtiles_batch_table.html
@@ -56,7 +56,7 @@
                     itowns.C3DTBatchTableHierarchyExtension });
 
             // Create the 3D Tiles layer
-            var $3dTilesLayerBTHierarchy = new itowns.ThreeDTilesLayer(
+            var $3dTilesLayerBTHierarchy = new itowns.OGC3DTilesLayer(
                 '3d-tiles-bt-hierarchy', {
                     name: 'BTHierarchy',
                     source: new itowns.C3DTilesSource({
